+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Documentation for public API
 - `row_range` range class to simplify use of `row_iterator`
 - Compile type requirements for `database::exec` callback
+- Overloads of `context::result_reference` that accept destructor function callback
 
 ### Changed
 - Defining SQLite configuration macros `SQLITE_OMIT_LOAD_EXTENSION`, `SQLITE_OMIT_PROGRESS_CALLBACK` and `SQLITE_OMIT_DECLTYPE`